requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
python-dotenv==0.19.0
pytest==6.2.5
mysql-connector-python==8.0.27
pymysql==1.0.2
Pillow==9.0.0
werkzeug==2.0.1